Refactoring to make list rendering a little faster
